Add files via upload
Hi Everyone,
This is Alec from the TON Foundation.
Please see the current version of our Liquidity Rewards program going live this month. We've listened to the feedback from the community and found a way to create more frequent payouts and simplify the program.

We'd love your feedback and any FAQ's that the community comes up with so we can help make the workings of this program more clear. We're also considering applying the TVL Open Challenge in this same form to a range of other verticals that can be measured on TVL

Thanks!